Add files via upload
Backtesting strategies of SP500 EFT Index, compared with the VIX Index with the SP500 Index being the benchmark. 
The strategy is to use volatility index as the basis of hold or sell. When VIX>30, sell and hold cash. When VIX<30, buy back and enter markets. Benchmark is the SP500 Index. 
I could not access a real-life trading API (very expensive), so i used daily data from YF from 2015 to now().